CSS image rewrite should always use forward slashes
The CSS image rewriting code was using node.js' path module, which has a
platform-specific path separator (e.g. "\" on Windows). Back slashes are
not valid for URLs, and so when a hashed URL is returned it should be
converted into a UNIX-style URL with forward slashes. Otherwise, images
referenced from CSS won't load.