Add support for customizing the renamed files pattern
Two new parameters are added which allow for customization of the
renamed file pattern:

-l, --hash-length, options.hashLength: allows specifying how many
characters of the hash to include

-f, --rename-format, options.renameFormat: specifies a template string
to use when renaming files

The default values are consistent with the previous implementation.
Overriding these values will allow you to customize how much of the hash
is used in the filename, and what the resulting renamed file looks like.