Add sourcemap-include-path option
Adds a new --sourcemap-include-path/sourcemapIncludePath option which is
a boolean. Default value is TRUE, and preserves existing behavior of
including the full absolute (virtual) path to the map file from the base
directory.

Setting the value to FALSE will cause the sourcemap to be written out as
relative to the source file without any path component.